+++ sync-for-expenses/src/sdk/models/shared/supplier.ts
@@ -3,62 +3,11 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { AccountingAddressType } from "./accountingaddresstype";
+import { Items } from "./items";
 import { Metadata } from "./metadata";
 import { SupplementalData } from "./supplementaldata";
 import { SupplierStatus } from "./supplierstatus";
 import { Expose, Type } from "class-transformer";
-
-export class SupplierAccountingAddress extends SpeakeasyBase {
-    /**
-     * City of the customer address.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "city" })
-    city?: string;
-
-    /**
-     * Country of the customer address.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "country" })
-    country?: string;
-
-    /**
-     * Line 1 of the customer address.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "line1" })
-    line1?: string;
-
-    /**
-     * Line 2 of the customer address.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "line2" })
-    line2?: string;
-
-    /**
-     * Postal code or zip code.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "postalCode" })
-    postalCode?: string;
-
-    /**
-     * Region of the customer address.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "region" })
-    region?: string;
-
-    /**
-     * The type of the address
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "type" })
-    type: AccountingAddressType;
-}
 
 /**
  * > View the coverage for suppliers in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=suppliers" target="_blank">Data coverage explorer</a>.
@@ -73,10 +22,10 @@ export class Supplier extends SpeakeasyBase {
     /**
      * An array of Addresses.
      */
-    @SpeakeasyMetadata({ elemType: SupplierAccountingAddress })
+    @SpeakeasyMetadata({ elemType: Items })
     @Expose({ name: "addresses" })
-    @Type(() => SupplierAccountingAddress)
-    addresses?: SupplierAccountingAddress[];
+    @Type(() => Items)
+    addresses?: Items[];
 
     /**
      * Name of the main contact for the supplier.